# docs/swagger/swagger.yaml
basePath: /
definitions:
  github_com_exnodes_hrm-api_internal_dto.AdminChangeEmailRequest:
    properties:
      new_email:
        type: string
    required:
    - new_email
    type: object
  github_com_exnodes_hrm-api_internal_dto.AdminUserPatch:
    properties:
      is_active:
        type: boolean
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.ChangeEmailRequest:
    properties:
      current_password:
        minLength: 1
        type: string
      new_email:
        type: string
    required:
    - current_password
    - new_email
    type: object
  github_com_exnodes_hrm-api_internal_dto.ChangePasswordRequest:
    properties:
      current_password:
        minLength: 1
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - new_password
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.DeleteUserRequest:
    properties:
      current_password:
        minLength: 1
        type: string
    required:
    - current_password
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.FcmTokenRequest:
    properties:
      device_id:
        type: string
      platform:
        enum:
        - android
        - ios
        - web
        - unknown
        type: string
      token:
        type: string
    required:
    - device_id
    - token
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.NotificationSettingsRequest:
    properties:
      notifications_enabled:
        type: boolean
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.RoleAssignmentRequest:
    properties:
      role_ids:
        items:
          type: string
        type: array
    required:
    - role_ids
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HRM API for Exnodes — see /docs/superpowers/specs/2026-05-15-go-migration-design.md
    for the migration plan.
  title: Exnodes HRM API
  version: 0.1.0
paths:
...
  /api/v1/users:
    get:
      description: Paginated list with optional email search and is_active filter.
        Each item embeds roles and the employee summary.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Substring match on email (ILIKE)
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user uuid
        in: path
        name: id
        required: true
        type: string
      - description: admin's current password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete a user account (admin reauth)
      tags:
      - users
    get:
      description: Returns the user with roles and embedded employee summary.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID (admin)
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: user uuid
        in: path
        name: id
        required: true
        type: string
      - description: fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.AdminUserPatch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Admin toggle is_active on a user
      tags:
      - users
  /api/v1/users/{id}/change-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: user uuid
        in: path
        name: id
        required: true
        type: string
      - description: new email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.AdminChangeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin changes a user''s email (employees parity #13)'
      tags:
      - users
  /api/v1/users/{id}/change-password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: user uuid
        in: path
        name: id
        required: true
        type: string
      - description: new password (current ignored)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Admin resets a user's password
      tags:
      - users
  /api/v1/users/{id}/roles:
    put:
      consumes:
      - application/json
      parameters:
      - description: user uuid
        in: path
        name: id
        required: true
        type: string
      - description: role ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.RoleAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign roles to a user (admin)
      tags:
      - users
  /api/v1/users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user (auth profile + embedded employee summary)
      tags:
      - users
  /api/v1/users/me/change-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: new email + current password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.ChangeEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change current user email
      tags:
      - users
  /api/v1/users/me/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change current user password
      tags:
      - users
  /api/v1/users/me/device-tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: device id + token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.FcmTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Register a device token for push notifications
      tags:
      - users
  /api/v1/users/me/device-tokens/{token}:
    delete:
      parameters:
      - description: FCM token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove a device token
      tags:
      - users
  /api/v1/users/me/notification-settings:
    patch:
      consumes:
      - application/json
      parameters:
      - description: settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.NotificationSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update push notification settings
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"